requirements.txt
# Streamlit & core packages
streamlit
pandas
openpyxl
pdfplumber

# Adobe PDF Services SDK
adobe-pdfservices-sdk==4.1.1


# LLMWhisperer SDK
requests
llmwhisperer-client
unstract-client

# LlamaIndex & LlamaExtract
llama-cloud-services

# Pydantic (for future schema-based extraction, optional but safe)
pydantic

# Other utilities
python-dotenv
